[feature-parallel-requests-future] Refactoring statische Methoden aus JenkinsClient in JobAbfrage umgezogen
